Adds the ability to completely hide data sources from users in AMScanView.
This is accomplished by using the Qt::ItemIsEnabled flag within AMScanSetModel. For AMScanView, this flag determines whether the data source is visible within plots, AND whether buttons to toggle that visibility should even be shown at all.  If the item is disabled, it is not visible, and there is no way (yet) for users to make it visible.  (This was based on the work Darren and I did together, but it turned out to be a bit more complicated because we initially only thought about it from the perspective of AMScanControllers; we forgot about having to also get the correct data sources hidden when re-loading from the database.)

You can configure all data sources like this in advance, using AMDataSource::setHiddenFromUsers(), or use the convenient overloaded versions of AMScan::addRawDataSource() and AMScan::addAnalyzedDataSource() which provide a way to set this and add the data source in one shot. This should allow updating existing code easily by adding two arguments where AMScanControllers create raw data sources.

This commit also adds the ability to pre-specify whether data sources should be visible in the multi-source plots, based on the AMDataSource::setVisibleInPlots() hint.

We now have two concepts for visibility:
- AMDataSource::hiddenFromUsers() <-> AMScanSetModel::flags() !ItemIsEnabled means hidden completely and cannot be made visible; there is no button shown to make it visible. In standard Qt views, the data source will be shown but grayed out like any disabled item.
- AMDataSource::visibleInPlots() <-> AMScanSetModel::data(Qt::CheckStateRole) indicates the user-controlled visibility state, which they are free to toggle even though programmers can pre-load a default.

New database required. For AMRawDataSources and AMAnalysisBlocks, the hiddenFromUsers() and visibleInPlots() hints are stored with the scan so they can be recalled when loading from the database.

This almost completes GH-175.  Darren, you'll want to go through your scan contrllers and make sure to use the new version of AMScan::addRawDataSource() on anything you want to hide, including the 0D data sources that were previously automatically hidden.  David, you'll find that analysis blocks are visibleInPlots() by default, but raw data source aren't; if you want to change any of those specifically in your scan controllers or importers, use the overloaded version of AMScan::addRawDataSource() and AMScan::addAnalyzedDataSource().