Adds rigorous checking for scans acquiring or already open when users try to open them out of the database.
For already-open, the user gets a warning and can choose to open a second copy (knowing the risks), or be taken to the already-open instance.

For currently acquiring, the user is not allowed to open again (since data will not be updating properly in the second copy). They can choose to be taken to the already-open scan, or just cancel.

Potential problem to keep our eyes on: if a scan controller crashes and doesn't remove the currentlyScanning column status for the scan, it will never be possible to open it (without manually digging into the database to change it.)

Closes #101.