Removes auto-tracking of the run date range based on the scans inside it, as well as the endDateTime() of a run, for database performance.
This might be a minor reduction in usability for users, who will now only see when the runs were created, instead of the date-range of the scans inside them. However, in most cases the creation time of the run will be close enough to the first scan inside it that most people won't notice the difference.

The motivation for removing this is for substantially-improved performance when saving scans in the database, as well as simplifying multithreading issues.