Adds a 10 second retry period to AMDatabase to avoid "Database is locked" errors.
Sqlite allows multiple connections from different threads, but locks the database during updates or selects. By default, an attempt to write while the database is locked will return a "Database is locked" error immediately.  This adds an automatic retry period of up to 10 seconds before failing. (Note: this may block or lock a thread for this long if it actually takes this long for a select or update to finish. Therefore, database calls should be organized when possible in multiple short queries unless a full transaction is needed. Selects should be closed as soon as possible. QSqlTableModel should not be used, since it leaves a select running the whole time it is open.)