Shortens length of time that SQL queries lock the database for across the whole program. Replaces QSqlTableModel with our own optimized, row-by-row loading, cached table model called AMQueryTableModel (since QSqlTableModel leaves a SELECT open as long as it exists, which does not play nice with SQLite's inter-thread database locking).
The new AMQueryTableModel only loads the row data as it is required, and caches all the values it does load.  Since each row is loaded in a separate SELECT, the lock-to-lock time is very short. (The longest SELECT happens when querying all the row ids that should be in the table according to the WHERE and ORDER BY clauses.)