Fixes a crash bug that was introduced into AMExportController when the usingScanObjects_ mode was added. Fixes #202.
It would access scanURLsToExport_(searchScanIndex++) even when there were no scanURLsToExport_, causing a crash in the non-scanURLsToExport_ mode.

This commit also adds some database import work in-progress, and also changes the behaviour of AMDatabase::select() to not execute the query.. You need to do this yourself now by calling q.exec() on the returned value. This allows binding of placeholder values before executing. (Updates all places that use this.)