Improves thread-safety of AMDbObject::storeToDb()'s thumbnail generation. Now objects can specify via virtual AMDbObject::shouldGenerateThumbnailsInSeparateThread() whether they can and want to do this in another thread.
Other-thread saving is disabled by default, so not all AMDbObject classes need to be re-entrant.  You have to intentionally enable it, which is turned on for just AMScan right now.  However: watch out: if you enabled this on composite classes that contain other classes as children: all of the composite classes also should be re-entrant in their loadFromDb(), thumbnail() and thumbnailCount() functions.