Fixes bug in REIXSAppController with respect to receiving scans from the AMScanControllerSupervisor::currentScanController().
This is one thing that should probably be moved into the framework (maybe the scan action?) so that if you forget to do it, you aren't leaking AMScans and also wondering why they aren't appearing. The new AMAppController has a nice way to receive scans and open them either in an existing editor, or a new editor... we just need to use this from scan action.