Fixes 3 crash bugs related to how AMBeamlineScanAction and AMScanControllerSupervisor delete the scan controller in edge cases and error situations.
At line 122 in AMBeamlineScanAction.cpp:
Used to be: if(!AMScanControllerSupervisor::scanControllerSupervisor()->setCurrentScanController(ctrl_) )

Problem: There are two ways for this function to fail. If setCurrentScanController() fails because the controller's scan is null (ie: not because there is already a valid currentScanController()), then this dialog will be shown when it shouldn't be.  Further, if the user chooses requestFail, then the ctrl_ will be deleted even though it's already been set as the current scan controller... So that when the next scan runs, setCurrentScanController() will crash when trying to delete the last currentScanController (because it's been deleted already, but not through deleteCurrentScanController().

At (old)line 167 in AMBeamlineScanAction.cpp:
Used to be: delete ctrl_;

Problem: At this point, the AMScanControllerSupervisor::currentScanController() was already set with the current ctrl_. Deleting it here but not deleting it through AMScanControllerSupervisor::deleteCurrentScanController() will cause a crash the next time a scan runs and this function calls setCurrentScanController(). That function will attempt to disconnect and delete the already-deleted controller through its non-zero but now-invalid pointer.

At (old)line 218 in AMBeamlineScanAction.cpp:
Same problem.