Fixes a bug in database registration system (AMDbObjectSupport) when registering a database after classes have already been registered in a previous database. Fixes #213.
Problem explanation:
The database system still supports classes sharing tables with other classes (although this might be recognized as a silly idea now.) Currently AMScan and all the scan subclasses share a table.  The problem was: When registering a database AFTER classess have already been registered, the system retroactively prepares the new database for all the classes that have been previously registered... But it was doing this in alphabetical order instead of registration order. In this case, it was trying to register AMFastScan before AMScan, even though AMFastScan counts on the AMScan_table existing already.

Fix description:
For now, it's been fixed to go through the existing classes in registration order.  For the future, we might want to consider getting rid of the table-sharing system altogether, which would simplify the design of the DB registration system a lot. This would require moving to a single AMScan class, which means finishing up the "complex" database upgrade system.

Note that it's still necessary, when choosing the order in which to register classes via AMDbObjectSupport::registerClass(), to make sure any table-sharing classes are registered AFTER their non-table-sharing buddy is registered.