Resolves a problem where AMControl could not described situations where the move PV had a different number of enums than the read PV.  Enables AMControl to expose separate sets of enum values for feedback (value()) and setpoint  (move()).  enumNames() now returns the same feedback enum values it always did, but moveEnumNames() now exists to return a different set of enums, if directed by the implementation using setMoveEnumStates().
Updates AMPVControl and AMPVwStatusControl to setMoveEnumStates() based on their writePV_.

Updates AMControlEditor to show the moveEnumNames() in the drop-down menu.

TODO: AMExtendedControlEditor.cpp should probably be updated too, to use moveEnumNames() instead of enumNames() on line 424... But this would possibly be a behaviour change for the SGM beamline: users might see a different set of enum choices than they were used to.  Leaving this alone for @davidChevrier to check out.