Revert "Removes "is scan acquiring check" from AMGenericScanEditor on scan close, since it never really belonged here, and now that we have the retain()/release() mechanism, scans can safely continue acquisition even if the editor is closed."
This reverts commit 7942a55f0d9ebaa3f208ed91cdf7f888fc922ec9.

Reason for reverting: I realized that this allows the possibility of accidentally closing the currently-acquiring scan, but then having no (current) way to re-open it.  Should probably discuss first with @davidChevrier and @dretrex.

In theory, we can now allow "opening" currently-acquiring scans from the database views... But instead of creating+loading the AMScan instance out of the database, we would need to:
- Find out a pointer to the currently-acquiring scan instance [using AMScanControllerSupervisor, or the workflow once we figure that out].
- Check if it's open in any editors already. [Not supported to have the same AMScan instance in multiple editors yet.]
- If not, open a new editor for it.