ADDED PVs to monitor mono angle motor and directky access to REIXS:energy
bl()->photonSource()->monoMirrorAngleStatus()
bl()->photonSource()->bipassEnergy()
REIXSBrokenMonoControl::onMonoMirrorAngleError(double error)

now on mono move, connection is made to monitor the mirror angle status
and on a move error the mono is stopped and the last move retriggered.

Initial testing succesful. It's a hack solution, but it'll do until the
AM framework decides how to better hand energy apps in general.

Closes Issue #680
2 hours