Altered:
- Order in which startup actions are performed in
AMDatamanAppController::startup() ~ Now checks if a backup is needed
before initializing for first time use. This means that if local
versions of files have been deleted, running AM will replace them,
instead of creating files as if it was a first time use, and then
writing over the remote files with the new blank ones.
- AMDirectorySynchronizerDialog::start() now displays a nice message to
the user if it encountered an error with the copy process before
starting.
- AMDirectorySynchronizer::start() now sets the percentage complete to
100% if it finds the files are already the same.