Added: - AMLightweightScanInfoFilterProxyModel ~ Needed to allow filtering of scans by row as a top level filter (on top of all the others)
Altered:
- Removed filtering by experiments from public AMScanDataView interface,
and calls to it in AMDatamanAppController. There doesn't seem to be any
reference to experiments. If this is used somewhere, maybe in other
beamlines, then I'll add it back
- AMLightweightScanInfoCollection ~ Now a QObject encapsulating a QList,
rather than just a subclass of QList. This allows for it to emit signals
when the data in it changes
- AMLightweightScanInfoCollection ~ A whole bunch of changes to cope
with changes to the data (adding/removal/updating of scans, updating of
runs and samples)
- AMLightweightScanInfoModel ~ Now listens for the signals emitted from
the AMLightweightScanInfoCollection it contains, and emits its own
signals to indicate that data is [about to be/has just been] changed.