Added: - AMScanTreeViewDelegate ~ The default item delegate for QTreeView behaved oddly when given strings which were very long (it set the height such that it had loads of room to display the long string, but then didn't display it). This class overrides that behaviour to give each row in the tree view a consistent height
Altered:
- AMScanDataView and AMLightweightScanInfoModel's methods of getting the
scan urls for selected items. Fixes a couple of bugs: If a thumbnail is
selected, it now adds the scan url for its parent as would be expected.
Also it now adds a scan only once, for instance in the circumstance
where a user has multiple thumbnails for the same scan selected when
they chose a context menu option.