Fixed IDEAS autoExport folder issue by disabling auto export.  Done.
modified:
IDEASXASScanConfiguration to include ```setAutoExportEnabled(false)```
in the constuctor and copy constructor.

Also snuck a ```setAutoExportEnabled(false)``` into the
REIXSXASScanConfiguration constructor as it was only in the copy
constructor (?).  I don't know what effect this woul dhave, if any, but
I choose to follow VESPERS venerable example and place it in both.

N.B. If anyone ever follows through on the comment in
```AMScanAction::autoExportScan()```: // This needs to be generalized so
the user can set it (on beamlines where this is acceptable), I may
revisit this apporach.  For now, this works.