Added: - AMScanThumbnailView::setSelectionBetween() ~ Sets all the items between the two passed model indices as selected using the passed selection command
Altered:
- AMScanThumbnailView::mouseReleaseEvent() ~ Now, if the shift button is
held down but the user is not dragging a box for the selection, the view
iterates backwards to the previously selected item, and then selects all
model indices between the two.

- AMScanThumbnailView::mousePressEvent() ~ No longer calls base version.
QAbstractItemView was doing too much logic on its own back, which was
interfering with the required logic that we need from the thumbnail
view. This means that:

Still ToDo:
- Need to rewrite double-click logic, as this was one of the things that
the QAbstractItemView was doing for us (amongst other more annoying
things).