Added: - A dead time view factor to change how the dead time buttons are arranged based on the number of elements there are.  It still defaults to the sqrt of the number of elements, but you can specify the number inside your detector view's constructor. - Added the threshold for each element to the BioXAS detector and view so we don't have to change them manually in the terminal and also started displaying the number of frames.
Altered:
- Changed the axis scale of the data sources so that they display properly.  Since the detector still had it's calibration, it turned out it has the same scaling of 10 eV/bin.

A general note, I don't listen to the threshold change since it is just meant to be a debgging tool for the time being.  I also need make sure (probably in the cleanup actions/state machine of the detector) that I clear 'ERASE' then 'UPDATE' the detector between acquisition's because otherwise the ring buffer is not cleared and you will get all kinds of crazy nonsense. #1178.