Removed several classes from bioxasMainCommon.pri: - BioXASMainMonochromatorCrystalChangeControl - BioXASMainMonochromatorCrystalChangeControlView - BioXASMainMonochromatorCrystalChangeView and replacing them with others that I've been working on alongside: - BioXASMainMonochromatorRegionControl - BioXASMainMonochromatorRegionControlView - BioXASMainMonochromatorRegionView
The RegionControl is a little different structurally from the CrystalChangeControl, it it a descendant of AMControl (for one) so it will be able to be edited and viewed in a similar way as the mono energy. It relates to the Monochromator class differently, in that there is no confusing 'two-way' relationship. The RegionControl is a member of BioXASMainMonochromator, and the mono class gives it all the necessary tools to perform it's job of managing the region (sets the appropriate controls). The mono class also doesn't manage any actions related to the crystal change, all have been moved to within the RegionControl. I think this makes the mono class much less cluttered.