The bulk of the refactor is complete, both Side and Main applications run and view the appropriate carbon filter farm information.
BioXASCarbonFilterFarmActuatorControl is a control that accounts for the status check after position moves. It's current value is the current actuator position.

BioXASCarbonFilterFarmActuatorWindowControl maps actuator positions to a 'window.' It's current value is the current window in the beam.

BioXASCarbonFilterFarmActuatorWindowFilterThicknessControl maps actuator windows to a filter thickness. It's current value is the filter thickness of the window in the beam.

BioXASCarbonFilterFarmFilterThicknessControl maps total filter thickness to thicknesses for each actuator. It's current value is the total filter thickness for windows that are in the beam.

Removed the beamline-specific filter farm controls.