Added:
AMScanThumbnailGrid::backgroundRect to serperate the background
rectangle generation from rectangle generation for model elements.

Moved:
AMScanThumbnailGridViewItemDelegate::paintItemBackground to public so it
can be called directly.

Altered:
In AMScanThumbnailGridView adjusted calls to create and draw thumbnail
background boxes outside of the methods used to render model elements
since the background is not explicity tied to the model. Previously the
serial index was used to draw the background box.