Modified:
BioXASPersistentView with a layout and box for a shutter control button
with indicator as it is setup in VESPERS per the issue suggestion.

Added:

BioXASPersistentView::onSOEShutterChanged as a slot to handle UI changes
when the shutter changes state. Other functionality is leveraged from
BioXASBeamline as SOE is instantiated and controlled there.

Notes:

Since the endstation shutter does not have a data class to itself I
added the view component directly to the persistent view. Alternatively
this could be seperated into its own view to be encapsulated better.