Modified:
AMControl::addChildControl had a logic bug in its conditional requiring
a child to exist in the children_ list before it could be added. I
suspect this would never allow a child control to be added. Since all
inheriting classes redefine the method this likely was not a serious
issue.

Undid my initial refactors as the whole hierarchy requires work which
will change my approach.