Modified:
AMChooseDataFolderDialog::getDataFolder()

Fixed another bug where if you switched from a data folder specified by a
short file path to a data folder specified by a full path the
remoteDataFolder .ini setting would still point to the previous data
folder.

To avoid this, the remoteDataFolder setting is deleted if the current
entry does not contain the data folder name for the selected project.