Added:
CLSControlsStatusView

A class to visualize control status information in the same style as the
CLSBeamlineStatusView. Currently the same the code but seperated to
allow for better generalization for any group of controls that may
benefit from being stylized that way.

CLSControlStatusButtonBar

Same idea as the CLSBeamlineStatusButtonBar but seperated to allow for
adjustment for handling new controls.