Modified:
Removed a check that did not make sense.

Adjusted AMGenericScanEditor to check for an empty, no selected file
case. As it would create a file at the directory above named
directoryName.ext which seems undesirable.

AMGenericScanEditor won't let you save with a file name input, which
needs to be fixed. Also cancelling saves the file, this may or may not be
desireable behiavor.