Continued conversion of AMScanViewSingleSpectrumView to use refactored code.
Commented out method bodies and referenced parent version of methods for
testing. Will remove the methods when I start to refactor
AMXRFDetailedDetectorView to insure method extension is not required for
the generalized versions.

Need to go over connections and socket methods to make sure that things
are joined properly in the new hierarchy.