Added:
A new export option in AMExporterOptionGeneralAscii containing a map of
data source names with thier precision values, along with getters and
setters.

Added an overloaded version of toString() to AMNumber to take a single
integer for precision while leaving old functionality intact.

Modified:

Started to make changes to exporters using AMExporterOptionGeneralAscii
features that output values with toString() to now specify the precision
for that datasource being handled.